backend/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
pandas==2.1.3
scikit-learn==1.3.2
djangorestframework==3.14.0
django-filter==23.5
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.0
setuptools==68.2.2
django-cors-headers==4.3.1
Pillow==10.3.0
faker==37.3.0
